一堆修改

+++ src/libc-base/alloc/mman.c
@@ -180,7 +180,7 @@ dlexport void *mman_aligned_alloc(mman_t man, size_t size, size_t align) {
     ptr = freelist_aligned_match(&man->large_blk, size, align);
   }
 
-  size_t offset = PADDING_UP(ptr, align) - (size_t)ptr;
+  size_t offset = PADDING_UP(ptr, align) - ptr;
   if (offset > 0) {
     void *new_ptr = blk_split(ptr, offset - 2 * sizeof(size_t));
     do_free(man, ptr);
